reorganized frontend js files
