+++ frontend/helpers/__tests__/clientUtils.test.js
