+++ frontend/static/js/ClientGame/ClientBoard.js
