+++ frontend/static/js/Command/__tests__/Command.test.js
