+++ frontend/static/js/GameDOM/GameView.js
